Start of code cleanup
Former-commit-id: 1b2c9b7658130fc66b2e06b4bc8cbca8a57863a4 [formerly 7eeb91c22611c44bdd7268a7bd4bb64e13d2e054]
Former-commit-id: a2d268d043d3163d0a1af8b621454766bd9af8b3

+++ src/librapid/array/functors/assignOp.hpp
@@ -181,7 +181,8 @@ void applyOp({0} **pointers, int64_t size) {{
 					blocksPerGrid	= 1;
 				} else {
 					threadsPerBlock = 512;
-					blocksPerGrid	= ceil(double(elems) / double(threadsPerBlock));
+					blocksPerGrid	= static_cast<int64_t>(
+						ceil(static_cast<double>(elems) / static_cast<double>(threadsPerBlock)));
 				}
 
 				dim3 grid(blocksPerGrid);
